+++ docs/changelog.md
@@ -1,6 +1,15 @@
 # Change log
 
+## 2025-03-08: Initial deployment scripts
+
+* Added initial setup script
+* Added initial teardown script
+
+I'll be using [PowerShell](https://learn.microsoft.com/en-us/powershell/scripting/learn/ps101/00-introduction?view=powershell-7.5) and the [Azure CLI](https://learn.microsoft.com/en-us/cli/azure/) to setup and teardown the resource group within Azure.
+
 ## 2025-03-01: Initial setup
 
 * Setup Github repository
 * Setup documentation
+
+I'll be using [GitHub](https://www.github.com) as a repository and [MkDocs](https://www.mkdocs.org/) for generating documentation.